refactor(cli): new internal cloudformation api module
