+++ packages/@aws-cdk/tmp-toolkit-helpers/src/api/resource-metadata/index.ts
@@ -1,0 +1,1 @@
+export * from './resource-metadata';
