+++ packages/aws-cdk/lib/api/deployments/index.ts
@@ -1,6 +1,4 @@
-export * from './cloudformation';
 export * from './deployments';
 export * from './deployment-result';
 export * from './deployment-method';
-export * from './nested-stack-helpers';
 export * from './asset-manifest-builder';
