+++ packages/aws-cdk/lib/api/environment/environment-access.ts
@@ -2,13 +2,13 @@ import type * as cxapi from '@aws-cdk/cx-api';
 import type { SDK } from '../aws-auth';
 import type { EnvironmentResources } from './environment-resources';
 import { EnvironmentResourcesRegistry } from './environment-resources';
+import type { StringWithoutPlaceholders } from './placeholders';
+import { replaceEnvPlaceholders } from './placeholders';
 import { ToolkitError } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api';
 import { IO, type IoHelper } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/private';
 import { formatErrorMessage } from '../../util';
 import type { CredentialsOptions, SdkForEnvironment, SdkProvider } from '../aws-auth/sdk-provider';
 import { Mode } from '../plugin/mode';
-import type { StringWithoutPlaceholders } from '../util/placeholders';
-import { replaceEnvPlaceholders } from '../util/placeholders';
 
 /**
  * Access particular AWS resources, based on information from the CX manifest
