+++ packages/aws-cdk/lib/api/environment/index.ts
@@ -1,2 +1,3 @@
 export * from './environment-access';
 export * from './environment-resources';
+export * from './placeholders';
