+++ packages/aws-cdk/lib/api/environment/placeholders.ts
