+++ packages/aws-cdk/lib/api/hotswap/ecs-services.ts
@@ -9,7 +9,7 @@ import {
 import { NonHotswappableReason, type ResourceChange } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/payloads/hotswap';
 import { lowerCaseFirstCharacter, transformObjectKeys } from '../../util';
 import type { SDK } from '../aws-auth';
-import type { EvaluateCloudFormationTemplate } from '../evaluate-cloudformation-template';
+import type { EvaluateCloudFormationTemplate } from '../cloudformation';
 
 const ECS_SERVICE_RESOURCE_TYPE = 'AWS::ECS::Service';
