+++ packages/aws-cdk/test/api/deployments/cloudformation-deployments.test.ts
@@ -12,7 +12,6 @@ import {
   CreateChangeSetCommand,
 } from '@aws-sdk/client-cloudformation';
 import { GetParameterCommand } from '@aws-sdk/client-ssm';
-import { CloudFormationStack, createChangeSet, Deployments } from '../../../lib/api/deployments';
 import { deployStack } from '../../../lib/api/deployments/deploy-stack';
 import { HotswapMode } from '../../../lib/api/hotswap/common';
 import { ToolkitInfo } from '../../../lib/api/toolkit-info';
@@ -28,6 +27,9 @@ import {
 } from '../../util/mock-sdk';
 import { FakeCloudformationStack } from '../_helpers/fake-cloudformation-stack';
 import { asIoHelper, TestIoHost } from '../../../../@aws-cdk/tmp-toolkit-helpers/src/api/io/private';
+import { Deployments } from '../../../lib/api/deployments';
+import { CloudFormationStack } from '../../../lib/api/cloudformation';
+import { createChangeSet } from '../../../lib/api/deployments/cfn-api';
 
 jest.mock('../../../lib/api/deployments/deploy-stack');
 jest.mock('../../../lib/api/deployments/asset-publishing');
